Integrate Metadata Aware Chunking with DBClient (#14)
* Pull out llm query + retriever query into method

* Add functionality in controller to extract references from retrieved docs and query retriever again

Note that the retriever seems to be performing a similarity match, which is clearly not enough.
It keeps failing. ReferenceExtractor has also been expanded to allow for multiple docs to be processed

* Test out functionality to add metadata to doc

* Pull headings from Document

We also attach the headings to Metadata as sections

* Expand TestFile script to take in multiple files

* Integrate metadata aware chunker with DBClient

The TestFile has been renamed. It has functions for chunking files into docs.
It also puts section metadata with each doc.
We replace the traditional Doc2txtloader with this method.

* bugfix: default section should be set to a string else Chroma can't read it

* Bugfix: allow system to be run in integrated fashion

* Commented out ref extractor and setup custom chunk splitter to mimic Doc2txtloader

* Test functionality to extract all relevant data from doc

docx.Document.paragraphs does not capture table data.

* Separate out full file chunking and current metadata aware chunking

* Build section based chunking ability with inchoate table parsing